Fix person calculation elsewhere too
