+++ posthog/temporal/ai/video_segment_clustering/activities/a5_generate_labels.py
@@ -7,16 +7,10 @@ import asyncio
 from datetime import datetime
 
 import structlog
-from asgiref.sync import sync_to_async
 from temporalio import activity
 
-from posthog.hogql import ast
-from posthog.hogql.parser import parse_select
-from posthog.hogql.query import execute_hogql_query
-
-from posthog.clickhouse.query_tagging import Product, tags_context
-from posthog.models.team import Team
 from posthog.temporal.ai.video_segment_clustering import constants
+from posthog.temporal.ai.video_segment_clustering.data import count_distinct_persons
 from posthog.temporal.ai.video_segment_clustering.labeling import generate_cluster_labels_llm
 from posthog.temporal.ai.video_segment_clustering.models import (
     ClusterContext,
@@ -27,37 +21,6 @@ from posthog.temporal.ai.video_segment_clustering.models import (
 )
 
 logger = structlog.get_logger(__name__)
-
-
-async def _count_distinct_persons(team_id: int, distinct_ids: list[str]) -> int:
-    """Count unique persons for a set of distinct_ids using HogQL."""
-    if not distinct_ids:
-        return 0
-
-    team = await Team.objects.aget(id=team_id)
-
-    @sync_to_async
-    def _execute_query():
-        with tags_context(product=Product.SESSION_SUMMARY):
-            return execute_hogql_query(
-                query_type="DistinctPersonCountForClustering",
-                query=parse_select(
-                    """
-                    SELECT COUNT(DISTINCT person_id)
-                    FROM person_distinct_ids
-                    WHERE distinct_id IN {distinct_ids}
-                    """
-                ),
-                placeholders={
-                    "distinct_ids": ast.Constant(value=distinct_ids),
-                },
-                team=team,
-            )
-
-    result = await _execute_query()
-    if result.results and len(result.results) > 0:
-        return result.results[0][0]
-    return 0
 
 
 async def _calculate_metrics_from_segments(team_id: int, segments: list[VideoSegmentMetadata]) -> dict:
@@ -71,7 +34,7 @@ async def _calculate_metrics_from_segments(team_id: int, segments: list[VideoSeg
 
     # Count unique persons via SQL (a person can have multiple distinct_ids)
     distinct_ids = [s.distinct_id for s in segments if s.distinct_id]
-    distinct_user_count = await _count_distinct_persons(team_id, distinct_ids)
+    distinct_user_count = await count_distinct_persons(team_id, distinct_ids)
 
     timestamps = []
     for s in segments:
